% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aopicspecies}
\alias{aopicspecies}
\title{Species List}
\usage{
aopicspecies()
}
\value{
Data frame with columns:
\itemize{
   \item \code{CommonName} — Common species name
   \item \code{Order} — Taxonomic order
   \item \code{Family} — Taxonomic family
   \item \code{LatinName} — Species Latin name
   \item \code{LensTransmissionInfoAvailable} — "Yes" if lens transmission data is available
   \item \code{OpsinSensitivityInfoAvailable} — "Yes" if any opsin data is available
}
}
\description{
Returns a table of species with information about lens transmission data 
and available opsin sensitivities.
}
\note{
Use the species common name (CommonName column) to specify data for 
that species when calling other functions in the library.
}
\examples{
aopicspecies()

}
