% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blambda}
\alias{blambda}
\title{Biological Light Detection for Light Pollution}
\usage{
blambda(power, wavelength, quantity = "luminous")
}
\arguments{
\item{power}{Numeric vector. Spectral irradiance in W/m²·nm or radiance 
in W/m²·sr·nm}

\item{wavelength}{Numeric vector. Wavelengths in nanometers (nm) 
corresponding to the power values.}

\item{quantity}{Character. Specifies the basis of the output system. Must 
be one of: "luminous", "radiant", or "photon". Default is "luminous".}
}
\value{
A data frame with the following columns: ("luminous" quantity = 
equivalent daylight illuminance in lux or luminance in cd/m²; "radiant" 
quantity = irradiance in W/m² or radiance in W/m²·sr; "photon" quantity = 
photon irradiance in log₁₀-scaled photons/cm²·s or photon radiance in 
log₁₀-scaled photons/cm²·s·sr).
\itemize{
   \item \code{Ba} — Animals  
   \item \code{Bp} — Plants and microorganisms  
   \item \code{Be} — All organisms  
}
}
\description{
Calculates luminous, radiant, or photon quantities 
using the \eqn{B_a(\lambda)}, \eqn{B_p(\lambda)}, and \eqn{B_e(\lambda)}
biological sensitivity functions. These functions represent a-opic biological
light detection for:
\itemize{
   \item all animals, and
   \item all plants and microbes, and
   \item all organisms, respectively.
}
}
\details{
\code{wavelength} must be defined in 1 nm increments.
}
\examples{
# Define wavelength range
wl <- 300:780
# Example spectral power distribution (W/m²·nm)
irradiance <- dnorm(wl, mean = 480, sd = 30)
# Biological light detection EDI values
exposure_animals  <- blambda(irradiance, wl)$Ba
exposure_plants   <- blambda(irradiance, wl)$Bp
exposure_general  <- blambda(irradiance, wl)$Be

}
