% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{govardovskii}
\alias{govardovskii}
\title{Govardovskii Visual Pigment Template}
\usage{
govardovskii(lmax, wavelength)
}
\arguments{
\item{lmax}{Numeric. Peak (\eqn{\lambda_{\max}}) of the photopigment in nm, 
measured without prereceptoral filtering.}

\item{wavelength}{Numeric vector. Wavelengths (nm) at which the action 
spectrum should be calculated.}
}
\value{
A numeric vector containing the relative photon sensitivity at each 
wavelength in \code{wavelength} (arbitrary units). No prereceptoral filtering 
is applied.
}
\description{
Computes the Govardovskii et al. (2000) universal visual pigment template,
returning the unfiltered photon action spectrum for a photopigment with a 
given peak wavelength (\eqn{\lambda_{\max}}). The function implements the 
Govardovskii et al. (2000) nomogram exactly as described in: Govardovskii 
et al. (2000). In search of the visual pigment template. Vis Neurosci. 
17(4), 509–528. doi: 10.1017/s0952523800174036.
}
\examples{
# Compute raw sensitivity of an opsin with λmax = 450 nm
wavelengths <- 300:800
aspecp <- govardovskii(450, wavelengths)

}
