% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{harmonisesens}
\alias{harmonisesens}
\title{Opsin sensitivity wavelength harmonizer}
\usage{
harmonisesens(Sens, wavelen)
}
\arguments{
\item{Sens}{existing sensitivity measurement data matrix with photon sensitivities and wavelengths (nm)}

\item{wavelen}{the wavelength range to extrapolate/restrict to}
}
\value{
adjusted sensitivtity curve
}
\description{
Extrapolate/adjust curve to match the wavelength range for species specific action spectra definition
}
\examples{
sens <- harmonisesens(data.frame(wavelen = 350:750, aspecp = govardovskii(450, 350:750)), 300:800);
}
