% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{harmonisetrans}
\alias{harmonisetrans}
\title{Prereceptoral filtering wavelength harmonizer}
\usage{
harmonisetrans(Pfilter, wavelen)
}
\arguments{
\item{Pfilter}{existing transmission measurement data matrix with transmissions (%) and wavelengths (nm)}

\item{wavelen}{the wavelength range to extrapolate/restrict to}
}
\value{
adjusted transmission data
}
\description{
Extrapolate/adjust species specific prereceptoral filtering data to match the wavelength range for action spectra definition
}
\examples{
trans <- harmonisetrans(data.frame(wavelen = 350:750, trans = seq(0,100,length.out=401)), 300:800);
}
