% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{photonenergy}
\alias{photonenergy}
\title{Photon Energy}
\usage{
photonenergy(wavelength)
}
\arguments{
\item{wavelength}{Numeric vector. Wavelengths in nanometers (nm).}
}
\value{
Numeric vector. Photon energy in Joules per photon.
}
\description{
Calculates the energy of a photon for given wavelengths in nanometers (nm),
based on Planck's equation: \eqn{E = hc / \lambda}.
}
\examples{
# Calculate photon energy for visible spectrum
wavelength <- 300:800
e <- photonenergy(wavelength)

}
