% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpeciesListData}
\alias{SpeciesListData}
\title{Species List Data}
\format{
A data frame with 4 variables:
\itemize{
  \item \code{order} Character. Taxonomic order of the species (e.g., "Primates").
  \item \code{family} Character. Taxonomic family of the species (e.g., "Hominidae").
  \item \code{species} Character. Common name of the species (e.g., "Human").
  \item \code{species_latin_name} Character. Scientific name of the species (e.g., "Homo sapiens").
}
}
\source{
Internal compiled dataset
}
\usage{
data(SpeciesListData)
}
\description{
This dataset provides taxonomic information for species included in the study.
It includes the taxonomic order and family, as well as the common and 
scientific names.
}
\examples{
data(SpeciesListData)
# List all species
unique(SpeciesListData$species)
}
\keyword{datasets}
