% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TransRefData}
\alias{TransRefData}
\title{Prereceptoral Lens Transmission Data}
\format{
A named list of species, each of which is a list with the following elements:
\itemize{
  \item \code{order} Character. Taxonomic order of the species (e.g., "Primates").
  \item \code{family} Character. Taxonomic family of the species (e.g., "Hominidae").
  \item \code{species} Character. Common name of the species (e.g., "Human").
  \item \code{species_latin_name} Character. Scientific name of the species (e.g., "Homo sapiens").
  \item \code{reference} Character. Source reference for the transmission data.
  \item \code{wavelen} Integer vector. Wavelength range (e.g., 300:800 nm).
  \item \code{trans} Numeric vector. Transmission values (\%) corresponding to each wavelength.
}
}
\source{
Internal compiled dataset
}
\usage{
data(TransRefData)
}
\description{
This dataset contains species-specific prereceptoral lens transmission curves.
Each element of the list corresponds to a species (e.g., "Human", "Mouse") 
and provides the reference transmission spectrum for that species.
}
\examples{
data(TransRefData)
# View species names
names(TransRefData)
# Plot human lens transmission
human_data <- TransRefData$Human
plot(human_data$wavelen, human_data$trans, type = "l",
     xlab = "Wavelength (nm)", ylab = "Transmission")
     
}
\keyword{datasets}
