% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VisualStandards}
\alias{VisualStandards}
\title{VisualStandards}
\format{
A list of datasets including D65, V(λ), Ba(λ), Bp(λ), Be(λ)
}
\source{
Internal compiled dataset
}
\usage{
data(VisualStandards)
}
\description{
Standard visual spectra used for photopic and biological light detection calculations.
}
\examples{
data(VisualStandards)

}
\keyword{datasets}
