% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vlambda}
\alias{vlambda}
\title{Illuminance/Luminance Calculator}
\usage{
vlambda(power, wavelength)
}
\arguments{
\item{power}{Numeric vector. Spectral irradiance in W/m²·nm or radiance in 
W/m²·sr·nm}

\item{wavelength}{Numeric vector. Wavelengths in nanometers (nm) 
corresponding to the power values.}
}
\value{
Numeric. Photopic illuminance (lux) or luminance (cd/m²).
}
\description{
Calculates photopic illuminance (lux) or luminance (cd/m²) based on the 
V(λ) photopic sensitivity function.
}
\details{
\code{wavelength} must be defined in 1 nm increments.
}
\examples{
# Define wavelength range
wl <- 300:780
# Example spectral power distribution in W/m²·nm
irradiance <- dnorm(wl, mean = 480, sd = 30)
# Calculate photopic illuminance
lux <- vlambda(irradiance, wl)
}
