% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aopicSpecies}
\alias{aopicSpecies}
\title{Species list}
\usage{
aopicSpecies()
}
\value{
species names
}
\description{
Show the list of species with available data of lens transmission and opsin sensitivity
}
\examples{
animals <- aopicSpecies();
}
