% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateaopicactionspec}
\alias{generateaopicactionspec}
\title{Alpha-opic action spectrum generator}
\usage{
generateaopicactionspec(opsin, lmax, pfilter)
}
\arguments{
\item{opsin}{target opsin name (One of the following options: 'Mel','Rod','Scone','Mcone','Lcone') or 'Photopic' human lux}

\item{lmax}{the lambda max of the opsin's photon sensitivity in the absence of preceptoral filtering. A string specifying a species specific work space in the subdirectory '/data' e.g. 'Mouse', or a numerical wavelength value specifying the lambda max to allow modelling}

\item{pfilter}{information about function for prereceptoral filtering. A string specifying a species specific work space in the subdirectory '/data' e.g. 'Mouse', zero for no prereceptoral filtering, or new transmission measurement data matrix with transmissions and wavelengths}
}
\value{
list of opsin, harmonized wavelength range, prereceptoral transmission, action spectrum, photon action spectrum, a-opic efficacy of luminous radiation for daylight. aspec = normalized a-opic action spectra for relative energy sensitivity as a function of wavelength. aspecp = normalized a-opic action spectra for relative photon sensitivity as a function of wavelength. kavD65 = a-opic efficacy of luminous radiation for daylight (D65; (W/lm)) for conversion to a-opic EDI. wavelen = the wavelength range over which aspec is defined (be default 300 to 780nm). trans = the prereceptoral filtering correction used to adjust the curves.
}
\description{
Generate a govardovski nomogram with corrections prereceptoral filtering
}
\examples{
as1 <- generateaopicactionspec('Photopic', 'Human', 'Human');
as2 <- generateaopicactionspec('Mel', 'Mouse', 'Mouse');
as3 <- generateaopicactionspec('Mel', 480, 'Mouse');
as4 <- generateaopicactionspec('Rod', 'Dog', data.frame(wavelen = 350:750, trans = seq(0,100,length.out=401)));
}
