% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{govardovskii}
\alias{govardovskii}
\title{Govardovskii nomogram}
\usage{
govardovskii(lmax, wavelen)
}
\arguments{
\item{lmax}{lambda max of the opsin's photon sensitivity in the absence of preceptoral filtering (nm)}

\item{wavelen}{a vector containing the wavelengths over which the nomogram is to be defined}
}
\value{
raw a-opic action spectra (relative photon sensitivity as a function of wavelength); arbitrary units, no correction for prereceptoral filtering
}
\description{
Calculate Govardovskii nomogram as per Govardovskii et al 2000 (Vis Neurosci. 2000 Jul-Aug;17(4):509-28. doi: 10.1017/s0952523800174036.)
}
\examples{
aspecp1 <- govardovskii(450, 300:800);
}
