% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{photonenergy}
\alias{photonenergy}
\title{Photon energy}
\usage{
photonenergy(wavelen)
}
\arguments{
\item{wavelen}{a vector containing the wavelengths}
}
\value{
photon energy (W.s/photon.nm)
}
\description{
Calculates the energy of a photon at a given wavelengths (nm)
}
\examples{
e <- photonenergy(300:800);
}
